[feature/post_e2e_testing] feat: implement Screen Robot pattern for Home, Post Detail, and Posts screens
